+++ pkgs/test/test/runner/runner_test.dart
@@ -374,7 +374,58 @@ $_usage''');
       await d.file('test.dart', _success).create();
       var test = await runDart(['test.dart']);
 
-      expect(test.stdout, emitsThrough(contains('All tests passed!')));
+      expect(
+        test.stdout,
+        allOf(
+          neverEmits(contains('success')),
+          emitsThrough(contains('All tests passed!')),
+        ),
+      );
+      await test.shouldExit(0);
+    });
+
+    test('directly with DART_TEST_REPORTER', () async {
+      await d.file('test.dart', _success).create();
+      var test = await runDart(
+        ['test.dart'],
+        environment: {'DART_TEST_REPORTER': 'expanded'},
+      );
+
+      expect(
+        test.stdout,
+        containsInOrder(['+0: success', '+1: All tests passed!']),
+      );
+      await test.shouldExit(0);
+    });
+
+    test('with DART_TEST_REPORTER from runner', () async {
+      await d.file('test.dart', _success).create();
+      var test = await runTest(
+        ['test.dart'],
+        environment: {'DART_TEST_REPORTER': 'expanded'},
+      );
+
+      expect(
+        test.stdout,
+        containsInOrder(['+0: success', '+1: All tests passed!']),
+      );
+      await test.shouldExit(0);
+    });
+
+    test('CLI flag overrides DART_TEST_REPORTER', () async {
+      await d.file('test.dart', _success).create();
+      var test = await runTest(
+        ['test.dart', '--reporter', 'failures-only'],
+        environment: {'DART_TEST_REPORTER': 'expanded'},
+      );
+
+      expect(
+        test.stdout,
+        allOf(
+          neverEmits(contains('success')),
+          emitsThrough(contains('All tests passed!')),
+        ),
+      );
       await test.shouldExit(0);
     });
 
@@ -444,7 +495,15 @@ $_usage''');
 
     test('directly', () async {
       var test = await runDart(['test.dart']);
-      expect(test.stdout, emitsThrough(contains('All tests passed!')));
+
+      expect(
+        test.stdout,
+        allOf(
+          neverEmits(anyElement(contains('success 1'))),
+          neverEmits(anyElement(contains('success 2'))),
+          emits(contains('All tests passed!')),
+        ),
+      );
       await test.shouldExit(0);
     });
   });