+++ pkgs/test_api/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.7.13-wip
+## 0.7.13
 
 * Add `Compiler.cli` (the native CLI compiler).
 * Expose several more backend APIs for use in `flutter_test`.